refactor: split backend code
